Make it build for Windows.

+++ src/thumbnail.c
@@ -252,7 +252,7 @@ bool nqiv_thumbnail_create(nqiv_image* image)
 		return false;
 	}
 	char mtime_string[NQIV_MTIME_STRLEN + 1] = {0};
-	snprintf(mtime_string, NQIV_MTIME_STRLEN, "%" PRIuMAX, stat_data.mtime);
+	snprintf(mtime_string, NQIV_MTIME_STRLEN, "%" PRIuMAX, (uintmax_t)stat_data.mtime);
 	char size_string[NQIV_SIZE_STRLEN + 1] = {0};
 	snprintf(size_string, NQIV_SIZE_STRLEN, "%zu", stat_data.size);
 	char width_string[NQIV_DIMENSION_STRLEN + 1] = {0};
